Network policy isn't actually selecting the pod (#5)
* feat: added network policy to allow traffic to the pod

The default hardened network policy only allows incoming traffic to the pod on the dedicated port it uses and nothing else

Signed-off-by: Bruno Bressi <bruno.bressi@telekom.de>

* fix: netpol syntax corrected

Signed-off-by: Bruno Bressi <bruno.bressi@telekom.de>

* chore: bumped to stable

Signed-off-by: Bruno Bressi <bruno.bressi@telekom.de>

* fix: corrected the navlinkswebhook network policy

This commit fixes the network policy label selectors. The current version uses the incorrect helm helmper function for the labels, and unfortunately this blocks communication to the webhook, as it is not selected by the network policy.

This must have the product of faulty testing on my side.

Signed-off-by: Bruno Bressi <bruno.bressi@telekom.de>

* chore: revert to new image

Signed-off-by: Bruno Bressi <bruno.bressi@telekom.de>

---------

Signed-off-by: Bruno Bressi <bruno.bressi@telekom.de>